Fix Graphics, UI, and Control Surfaces
- Fixed `web_ui/index.html` and `web_ui/app.js` to pass all Playwright verification tests.
- Fixed XSS vulnerability in window titles by safely escaping HTML characters in `spawnApp`.
- Corrected search debouncing by implementing `debouncedFilterApps` with `setTimeout` logic.
- Resolved `requestAnimationFrame` bugs for slider visuals by scoping logic accurately within `DOMContentLoaded`.
- Handled mock UI gaps for the empty state grid, and notifications to ensure UI surface flows complete successfully.
- Aligned Playwright test scripts to reference correct `.dock-item` classes instead of outdated `.icon` selectors.
- Fixed unused variable warnings in syscall stubs and cast issues in syscall tables that broke the CI build.

+++ kernel/fs/devfs.c
@@ -105,7 +105,7 @@ static uint32_t dev_binder_write(fs_node_t *node, uint32_t offset, uint32_t size
 
 static int dev_binder_ioctl(fs_node_t *node, int request, void *arg) {
     (void)node;
-    if (request == BINDER_VERSION_IOWR) {
+    if ((unsigned int)request == (unsigned int)BINDER_VERSION_IOWR) {
         if (!arg) return -1;
         struct binder_version *ver = (struct binder_version*)arg;
         ver->protocol_version = 8; /* Current Binder protocol version */
